Disable the persistent model cache by default
This feature, while making builds faster, is unstable and produces a
lot of hard to diagnose issues. I am disabling this for now. It can be
enabled by setting the environment variable ENABLE_PERSISTENT_MODEL_CACHE
to 1 manually. We can revisit this in the future if we want to
re-enable.