Moving discriminant parameters to component base
By moving the discriminant to the base type we ensure that the component
model always matches the implementation. When the discriminant is
reflected in the implementation, it was up to the user to keep the model
in sync with the code, which is far from ideal. This commit fixes that
issue.