Tweak GPR files so they work with gnatsas
For some reason, gnatsas requires that abstract project files declare an
empty source list, where as gprbuld has never required that. This commit
applies that change to abstract .gpr files supplied with Adamant and
replaces references to codepeer in the .gpr files with gnatsas, the new
name for the analysis suite.