Make `activate` install project py dependencies
This modification to `activate` will cause it to look in the env/
directory in any adamant projects in the path. If a requirements*.txt
file is found in that directory, it will attempt to pip install it. In
this way, projects can declare their own specific python dependencies,
that will then be automatically installed by Adamant along with the
Adamant-specific dependencies.