Update C packed record conversions
This allows the record conversions to work for types that previously
failed compilation checks due to alignment constraints of aliased
fields. This allows any packed record to produce a valid U_C type with
conversion to U. In addition, new conversion functions have been added
to convert directly from U_C to T and/or T_Le and back.