Fix for linking to unbuilt C/C++ objects
Linking to C/C++ would fail when only a portion of the C/C++ source
files in a directory are dependencies of the currently compiling binary.
Only some of the files would be compiled into objects, but GPRBuild (by
default) tries to link to all C/C++ objects found in a directory in the
Source_Dirs. To work around this, at link time, we now pass a list of
files to exclude from the build, which includes C/C++ files that exist
in the Source_Dirs, but are not dependencies of the currently compiling
binary.