Split components declaration into its own package
This provides a dedicated package to store assembly component
declaration, which also includes tasks and interrupt declaration. The
name of this file is <assembly>_components.ads.

By removing these from the overall autocoded <assembly>.ads we
resolve a circular depdendency problem that can sometime arise when a
generic component declares types that then need to be provided to it
in its Init.  In this case, a package that instantiates these types
(ie. something like assembly_registers.ads) would depend on the
assembly .ads. The assembly .adb would, in turn, depend on that
package, creating a circular dependency. Now, the package can
selectively import only the <assembly>_components.ads file, removing
its dependency on the entire assembly .ads, and thus removing the
circular dependency.