Add epsilon based assertions for packed types
that include one or many floating point values. When using the assertion
package for a packed record or array that includes one or more floating
point types, you can provide a single "Epsilon" argument as part of the
assertion, which is used to compare all floating point types inside of
the packed record or array. Non floating point types are still compared
for exact equality (or inequality) as before. By default, if not
specified, the value of Epsilon will be 0.0, resulting in the same
behavior as before this feature existed.