Ensure validation before parameter update
in the Parameters component. Prior to this commit, the "update"
operation was sent to all associated components during a table update
even if an entry in that table failed to validate. This behavior has now
been changed to not send the update operation (or subsequent table dump)
if the table failed to validate.